+++ ewoh-spark-app/server/app.module.ts
@@ -32,6 +32,8 @@ import { AuditModule } from './modules/audit/audit.module';
 import { OperationsModule } from './modules/operations/operations.module';
 import { ParametersModule } from './modules/parameters/parameters.module';
 import { AasModule } from './modules/aas/aas.module';
+import { TracingModule } from './modules/tracing/tracing.module';
+import { TracingInterceptor } from './modules/tracing/tracing.interceptor';
 
 @Module({
   imports: [
@@ -65,6 +67,7 @@ import { AasModule } from './modules/aas/aas.module';
     OperationsModule,
     ParametersModule,
     AasModule,
+    TracingModule,
     // ====== @route-section: business-modules END ======
 
     // ⚠️ @route-order: last